feat: provide backup and restore process
The container mounts are backed up, along with the LocalSettings.php, in a zip file, which is then
uploaded to the S3 bucket.

The restore process just involves extracting that zip file and putting the data back in the location
that the containers expect it to be.

I tested it and it works ok.

There is another restore playbook that is introduced, which has lots of duplication with the
preinstall playbook, but it doesn't really matter too much. This process is not expected to change
often.